String/0030-substring-with-concatenation-of-all-words/solution_test.go
/*
 * Copyright (c) 2020.  beihai@wingsxdu.com. All right reserved
 * ProjectName: $LeetCode
 * LastModified: 2020/3/16 下午9:45
 * Author: beihai
 */

package _030_substring_with_concatenation_of_all_words

import (
	"reflect"
	"testing"
)

func Test_findSubstring(t *testing.T) {
	type args struct {
		s     string
		words []string
	}
	tests := []struct {
		name string
		args args
		want []int
	}{
		{"test1", args{"barfoothefoobarman", []string{"foo", "bar"}}, []int{0, 9}},
		{"test2", args{"wordgoodgoodgoodbestword", []string{"word", "good", "best", "word"}}, []int{}},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := findSubstring(tt.args.s, tt.args.words); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("findSubstring() = %v, want %v", got, tt.want)
			}
		})
	}
}
